requirements.txt
umap-learn
visdom
librosa==0.8.1
matplotlib>=3.3.0
numpy==1.19.3; platform_system == "Windows"
numpy==1.19.4; platform_system != "Windows"
scipy>=1.0.0
tqdm
sounddevice
SoundFile
Unidecode
inflect
PyQt5
multiprocess
numba
webrtcvad; platform_system != "Windows"
pypinyin
flask
flask_wtf
flask_cors==3.0.10
gevent==21.8.0
flask_restx
tensorboard==1.15
streamlit==1.8.0
PyYAML==5.4.1
torch_complex
espnet
PyWavelets
monotonic-align==0.0.3
transformers==4.26.0
fastapi
loguru
typer[all]
click==8.0.4
